src/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
certifi==2020.12.5
    # via requests
cffi==1.14.5
    # via pynacl
chardet==4.0.0
    # via requests
deprecated==1.2.12
    # via pygithub
idna==2.10
    # via requests
pycparser==2.20
    # via cffi
pygithub==1.55
    # via -r requirements.in
pyjwt==2.1.0
    # via pygithub
pynacl==1.4.0
    # via pygithub
requests==2.25.1
    # via pygithub
six==1.15.0
    # via pynacl
urllib3==1.26.4
    # via requests
wrapt==1.12.1
    # via deprecated
